added wav2vec2 to onnx_model_bart
